// src/app/models/cliente.model.ts
export interface Cliente {
    idCliente?: string;
    nombre: string;
    celular: string;
    hora: string;
    tipo: string;
    descripcion: string;
    nivelAgua: number;
    fecha: string;
    mesaRegalos: boolean;
    colorSobremantel: string;
    capacidadPersonas: number;
    brincolin: boolean;
    precio: number;
    anticipo: number;
    tipoPago: string;
    totalPagar: number;
    restoPagar: number;
}